requirements.txt
websocket-client
slacker
eunjeon